requirements.txt
GitPython==3.1.47
PyYAML==6.0.3
yara-python==4.5.4
plyara==2.2.8
pySigma==1.3.3
